+++ sdk/formrecognizer/ai-form-recognizer/src/lro/analyze.ts
@@ -7,24 +7,33 @@ import {
   AnalyzeResult as GeneratedAnalyzeResult,
   AnalyzeResultOperation,
   AnalyzeResultOperationStatus as AnalyzeOperationStatus,
-  BoundingRegion,
-  Document as GeneratedDocument,
-  DocumentEntity,
-  DocumentKeyValuePair,
-  DocumentPage as GeneratedDocumentPage,
-  DocumentLine as GeneratedDocumentLine,
-  DocumentSelectionMark,
+  DocumentLanguage,
   DocumentSpan,
   DocumentStyle,
-  DocumentTable,
-  DocumentWord,
-  LengthUnit,
-  DocumentLanguage,
-  DocumentParagraph,
 } from "../generated";
 import { DocumentField, toAnalyzedDocumentFieldsFromGenerated } from "../models/fields";
 import { FormRecognizerApiVersion, PollerOptions } from "../options";
 import { AnalyzeDocumentOptions } from "../options/AnalyzeDocumentsOptions";
+import {
+  toBoundingPolygon,
+  toBoundingRegions,
+  toDocumentTableFromGenerated,
+  toKeyValuePairFromGenerated,
+} from "../transforms/polygon";
+import {
+  BoundingRegion,
+  DocumentEntity,
+  DocumentTable,
+  DocumentKeyValuePair,
+  DocumentPage,
+  DocumentLine,
+  DocumentParagraph,
+} from "../models/documentElements";
+import {
+  Document as GeneratedDocument,
+  DocumentPage as GeneratedDocumentPage,
+  DocumentLine as GeneratedDocumentLine,
+} from "../generated";
 
 /**
  * A request input that can be uploaded as binary data to the Form Recognizer service. Form Recognizer treats `string`
@@ -81,6 +90,7 @@ export interface AnalyzedDocument {
 export function toAnalyzedDocumentFromGenerated(document: GeneratedDocument): AnalyzedDocument {
   return {
     ...document,
+    boundingRegions: toBoundingRegions(document.boundingRegions),
     fields: toAnalyzedDocumentFieldsFromGenerated(document.fields ?? {}),
   };
 }
@@ -153,109 +163,6 @@ export interface AnalyzeResult<Document = AnalyzedDocument> extends AnalyzeResul
 }
 
 /**
- * A page within an analysis result.
- */
-export interface DocumentPage {
-  /**
-   * The kind of page. One of:
-   *
-   * - "document"
-   * - "sheet"
-   * - "slide"
-   * - "image"
-   *
-   * This list is non-exhaustive, and new values may be introduced in future versions of the Form Recognizer service.
-   */
-  kind: string;
-  /**
-   * 1-based page number in the input document.
-   */
-  pageNumber: number;
-
-  /**
-   * The general orientation of the content in clockwise direction, measured in degrees between (-180, 180].
-   */
-  angle: number;
-
-  /**
-   * The width of the image/PDF in pixels/inches, respectively.
-   */
-  width: number;
-
-  /**
-   * The height of the image/PDF in pixels/inches, respectively.
-   */
-  height: number;
-
-  /**
-   * The unit used by the width, height, and boundingBox properties. For images, the unit is "pixel". For PDF, the unit is "inch".
-   */
-  unit: LengthUnit;
-
-  /**
-   * Location of the page in the reading order concatenated content.
-   */
-  spans: DocumentSpan[];
-
-  /**
-   * Extracted words from the page.
-   */
-  words: DocumentWord[];
-
-  /**
-   * Extracted selection marks from the page.
-   */
-  selectionMarks?: DocumentSelectionMark[];
-
-  /**
-   * Extracted lines from the page, potentially containing both textual and visual elements.
-   */
-  lines?: DocumentLine[];
-}
-
-/**
- * Convert a REST-level DocumentPage into a convenience layer version.
- *
- * @internal
- * @param generated - a REST-level DocumentPage.
- * @returns
- */
-export function toDocumentPageFromGenerated(generated: GeneratedDocumentPage): DocumentPage {
-  // We will just overwrite the `lines` property with the transformed one rather than create a new object.
-  if (generated.lines) {
-    generated.lines = generated.lines.map((line) => toDocumentLineFromGenerated(line, generated));
-  }
-  return generated as DocumentPage;
-}
-
-/**
- * A line of adjacent content elements on a page.
- */
-export interface DocumentLine {
-  /**
-   * Concatenated content of the contained elements in reading order.
-   */
-  content: string;
-
-  /**
-   * Bounding box of the line.
-   */
-  boundingBox?: number[];
-
-  /**
-   * Location of the line in the reading order concatenated content.
-   */
-  spans: DocumentSpan[];
-
-  /**
-   * Compute the `DocumentWord`s that are related to this line.
-   *
-   * This function produces a lazy iterator that will yield one word before computing the next.
-   */
-  words: () => IterableIterator<DocumentWord>;
-}
-
-/**
  * Tests if one span contains another, by testing that the outer span starts before or at the same character as the
  * inner span, and that the end position of the outer span is greater than or equal to the end position of the inner
  * span.
@@ -282,12 +189,47 @@ function* empty(): Generator<never> {
  * @param items - the items to iterate over
  * @param idx - the index of the first item to begin iterating from
  */
-function* iterFrom<T>(items: T[], idx: number): Generator<T> {
+export function* iterFrom<T>(items: T[], idx: number): Generator<T> {
   let i = idx;
 
   while (i < items.length) {
     yield items[i++];
   }
+}
+
+export function toDocumentLineFromGenerated(
+  generated: GeneratedDocumentLine,
+  page: GeneratedDocumentPage
+): DocumentLine {
+  (generated as DocumentLine).words = () =>
+    fastGetChildren(
+      iterFrom(generated.spans, 0),
+      page.words.map((word) => {
+        return { ...word, polygon: toBoundingPolygon(word.polygon) };
+      })
+    );
+
+  Object.defineProperty(generated, "words", {
+    enumerable: false,
+  });
+
+  return generated as DocumentLine;
+}
+
+export function toDocumentPageFromGenerated(generated: GeneratedDocumentPage): DocumentPage {
+  return {
+    ...generated,
+    lines: generated.lines?.map((line) => toDocumentLineFromGenerated(line, generated)),
+    selectionMarks: generated.selectionMarks?.map((mark) => {
+      return { ...mark, polygon: toBoundingPolygon(mark.polygon) };
+    }),
+    words: generated.words.map((word) => {
+      return { ...word, polygon: toBoundingPolygon(word.polygon) };
+    }),
+    images: generated.images?.map((image) => {
+      return { ...image, polygon: toBoundingPolygon(image.polygon) };
+    }),
+  };
 }
 
 /**
@@ -374,27 +316,6 @@ export function* fastGetChildren<Spanned extends { span: DocumentSpan }>(
 }
 
 /**
- * Transforms a REST-level document line into a convenience layer version.
- *
- * @param generated - a REST-level DocumentLine
- * @param page - the page where the DocumentLine appeared
- * @returns a convenience layer DocumentLine
- */
-function toDocumentLineFromGenerated(
-  generated: GeneratedDocumentLine,
-  page: GeneratedDocumentPage
-): DocumentLine {
-  (generated as DocumentLine).words = () =>
-    fastGetChildren(iterFrom(generated.spans, 0), page.words);
-
-  Object.defineProperty(generated, "words", {
-    enumerable: false,
-  });
-
-  return generated as DocumentLine;
-}
-
-/**
  * The state of an analysis operation, which will eventually produce the result type that corresponds to the model.
  */
 export interface DocumentAnalysisPollOperationState<Result = AnalyzeResult<AnalyzedDocument>>
@@ -455,13 +376,16 @@ export function toAnalyzeResultFromGenerated<
     modelId: result.modelId,
     content: result.content,
     pages: result.pages.map((page) => toDocumentPageFromGenerated(page)),
-    tables: result.tables ?? [],
-    keyValuePairs: result.keyValuePairs ?? [],
-    entities: result.entities ?? [],
+    tables: result.tables?.map((table) => toDocumentTableFromGenerated(table)) ?? [],
+    keyValuePairs: result.keyValuePairs?.map((pair) => toKeyValuePairFromGenerated(pair)) ?? [],
+    entities:
+      result.entities?.map((entity) => {
+        return { ...entity, boundingRegions: toBoundingRegions(entity.boundingRegions) };
+      }) ?? [],
     languages: result.languages ?? [],
     styles: result.styles ?? [],
     documents: (result.documents?.map((doc) => mapDocuments(doc)) as Document[]) ?? [],
-    paragraphs: result.paragraphs ?? [],
+    paragraphs: result.paragraphs?.map(para => { return { ...para, boundingRegions: toBoundingRegions(para.boundingRegions) } }) ?? [],
   };
 }
 
@@ -475,7 +399,7 @@ export interface AnalysisOperationDefinition<Result = AnalyzeResult> {
   transformResult: (primitiveResult: GeneratedAnalyzeResult) => Result;
   initialModelId: string;
   options: PollerOptions<DocumentAnalysisPollOperationState<Result>> &
-    AnalyzeDocumentOptions<Result>;
+  AnalyzeDocumentOptions<Result>;
 }
 
 /**