+++ sdk/formrecognizer/ai-form-recognizer/src/models/index.ts
@@ -22,3 +22,20 @@ export {
   DocumentCurrencyField,
   DocumentField,
 } from "./fields";
+export {
+  BoundingRegion,
+  DocumentEntity,
+  DocumentImage,
+  DocumentKeyValueElement,
+  DocumentKeyValuePair,
+  DocumentLine,
+  DocumentParagraph,
+  DocumentSelectionMark,
+  DocumentTable,
+  DocumentTableCaption,
+  DocumentTableCell,
+  DocumentTableFootnote,
+  DocumentWord,
+  HasBoundingPolygon,
+  DocumentPage,
+} from "./documentElements";