+++ sdk/formrecognizer/ai-form-recognizer/CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Breaking Changes
 
 - Renamed the `beginCopyModel` method of `DocumentModelAdministrationClient` to `beginCopyModelTo`. [#20775](https://github.com/Azure/azure-sdk-for-js/pull/20775)
+- Renamed `BoundingRegion#boundingBox` to `BoundingRegion#polygon`, as the service may now provide arbitrary, polygonal bounding areas rather than just rectangles.
 
 ### Bugs Fixed